+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 #### Added
 - Tooltips for the 'is this appearance collected' icons on the model viewer's transmog set panel.
 
+#### Removed
+- Removed the 'transform' tab on the model viewer.
+    - I felt this tab wasn't necessary, and it probably didn't even work.
+
 #### Fixed
 - Fixed an error that would be thrown if a transmog set contained an invalid source ID.